Split rcom.py to per-vendor structure.
Split rcom.py to:
rat-com which is an executable file and entry-point for the user;
rcom.py which is a common base for all vendor-specific functions;
cisco.py which is a Cisco IOS supporting class;
procurve.py which is a HP ProCurve supporting class.

Fix rlist.py to support enabled/disabled flag and move the executable
code to rat-list.